# feat: Add incremental flash programming for faster embedded development iterations