--- a/probe-rs-tools/src/bin/probe-rs/util/flash.rs
+++ b/probe-rs-tools/src/bin/probe-rs/util/flash.rs
@@ -32,10 +32,11 @@ pub fn run_flash_download(
     let mut options = DownloadOptions::default();
     options.keep_unwritten_bytes = download_options.restore_unwritten;
     options.dry_run = probe_options.dry_run();
-    options.do_chip_erase = do_chip_erase;
+    options.do_chip_erase = do_chip_erase && !download_options.incremental; // Disable chip erase in incremental mode
     options.disable_double_buffering = download_options.disable_double_buffering;
     options.verify = download_options.verify;
     options.preverify = download_options.preverify;
+    options.incremental = download_options.incremental;
 
     let flash_layout_output_path = download_options.flash_layout_output_path.clone();
 
@@ -136,6 +137,8 @@ impl ProgressBars {
                 Operation::Fill => "Reading flash",
                 Operation::Program => "Programming",
                 Operation::Verify => "Verifying",
+                Operation::Crc32Verify => "Hash Check",
+                Operation::IncrementalProgram => "Programming",
             };
             ProgressBarGroup::new(format!("{message:>13}"))
         })