@fdomke 2025-07-24
>  ### Slicing-by-4 CRC32
>  - **Purpose**: High-performance CRC32 implementation for flash verification
>  - **Performance**: ~600-800 KB/s on RP2040 @ 125MHz
>  - **Memory**: 4KB lookup tables + ~400 bytes code  
>  - **Compatibility**: Works on all Cortex-M targets with >8KB SRAM
>  - **Optimization**: Processes 4 bytes simultaneously with aggressive loop unrolling

This seems overly a.) complex and b.) slow. The naive, table-driven implementation of crc32 [boils down](https://en.wikipedia.org/wiki/Computation_of_cyclic_redundancy_checks#CRC-32_example), when compiled with rustc,  to an inner loop of 11 instructions and a single 32-bit load from a 1kbyte table. I measured Rust's `crc` crate with `CRC_32_CKSUM` performance on a RP2040 (also at 125 MHz) and it gives me roughly 8.8 Mbyte/s. (I did not measure your code.)

Unfortunately
```
     ├── crc32_slice4_optimized.s        # Assembly source
 ```
is missing from the PR, but I don't think any of the loop-unrolling or slice-by-n would significantly speed up the flash process.

I looked at the disassembly and I think from just an instructions/byte ratio it's not better. Slice-by-N is mostly useful on memory-bandwidth constrained CPUs, which (especially lower-end) M cores are typically not.