@xobs 2025-08-13
I was considering this, and my approach would have been to inject a very simple CRC32 algorithm. One that didn't require any lookup tables, and was as small as possible.

The smallest I've been able to make in the last few minutes is:

```cpp
unsigned int crc32(const unsigned char *message, unsigned int len) {
   int j;
   const unsigned char *end = message + len;
   unsigned int crc;

   crc = 0xFFFFFFFF;
   while (message != end) {
      crc = crc ^ *message++;
      for (j = 7; j >= 0; j--) {
         crc = (crc >> 1) ^ (0xEDB88320 & -(crc & 1));
      }
   }
   return ~crc;
}
```

And in Rust:

```
pub fn crc32(message: &[u8]) -> u32{
   let mut crc = 0xFFFFFFFF;
   for byte in message {
      crc = crc ^ (*byte as u32);
      for _ in 0..=7 {
         let mask = (crc & 1)*u32::MAX;
         crc = (crc >> 1) ^ (0xEDB88320 & mask);
      }
   }
   !crc
}
```

Here's the Godbolt comparison, along with a Rust implementation that is unoptimized and is larger: https://godbolt.org/z/nrbE8qoxj

The crc32 C binary comes out to 52 bytes for M0, which should work on any ARM target. When compiled for rv32 it comes out to 80 bytes (58 if you build with compressed instructions).

I think it would be interesting to measure the speedup of the unrolled version versus the small-but-simple version.