@andrewleech 2025-08-21
# PR Update

**Major Update: Complete architectural rebuild with enhanced build system**

I've completely rebuilt this implementation with a focus on maintainability, shared configuration, and automatic validation. This version addresses all previous architectural concerns and adds significant build system improvements.

**Key Changes:**

**Flash algorithm ownership** (@bugadani's main requirement): The flash algorithm is now in charge of all target-side functionality. No more parallel systems or competing with existing pre-verification.

**Proper Rust project** (@Yatekii): Replaced Makefiles/shell scripts with a proper `crc32_algorithms/` Cargo workspace. Full source code included, no mystery binaries.

**Optimized CRC algorithm** (@fdomke, @Yatekii): Now using `crcxx` with CRC32_BZIP2 for both host and embedded code. Shared configuration ensures host/target consistency.

**RP2040 rescue mode compatibility**: Fixed the fundamental issue where this completely failed on RP2040 rescue mode boot sequences. Now works reliably across all RP2040 states.

**Shared configuration architecture**: Single source of truth for CRC algorithm across embedded firmware, host verification, and build metadata.

**Automatic offset detection**: Eliminates dangerous hardcoded function offsets using objdump symbol analysis.

**Technical Implementation:**

- **Flash algorithm-owned sequencing**: Proper init→uninit→CRC32→init lifecycle management
- **Shared CRC configuration**: `crc32_algorithms/src/crc_config.rs` provides single source of truth
  - Embedded firmware: `use probe_rs_crc32_builder::crc_config`
  - Host verification: Same shared config in `flasher.rs`
  - Build metadata: Automatic algorithm name generation
- **Automatic offset detection**: `objdump -t` symbol analysis eliminates hardcoded 0x00000008
- **Enhanced xtask system**: `cargo xtask build-crc32` with automatic validation
- **Clean dependency management**: Proper Cargo workspace integration
- **RP2040 XIP-aware state management**: Handles execute-in-place flash without corruption

**Performance Results (RP2040 @ 125MHz):**
- **CRC32_BZIP2 optimization**: 108.2ms vs 115.9ms per sector
- First flash: ~8-9s (includes CRC32 verification)
- Subsequent small changes: ~3-5s (96% reduction in flash operations)
- No changes: ~2-3s (pure verification)

**Testing:**
- RP2040 rescue dp mode compatibility (was completely broken before)
- STM32 cross-platform validation
- i.MX RT ARM Cortex-M7 validation
- CRC32_BZIP2 performance validation
- Build system integration with automatic binary validation
- Shared configuration prevents algorithm sync issues
- Automatic offset detection eliminates hardcoded maintenance burden

**Build System Enhancements:**
- **Maintainable**: Algorithm changes require only one edit in `crc_config.rs`
- **Type-safe**: Rust compiler ensures all components use same algorithm
- **Self-validating**: Automatic function offset detection from ELF symbols
- **Clean architecture**: Clear separation between embedded, host, and build systems

This addresses the feedback raised and adds significant maintainability improvements. 
The architecture is now cleaner, the RP2040 issues are resolved, and the build system prevents common maintenance pitfalls.